Refactoring structure of customer module Part 1
