--- src/module/customer/adapter/external/api/remove.customer.api.adapter.ts
@@ -1,31 +1,0 @@
-import {Injectable} from '@angular/core';
-import {customerEndpointEnum} from "@customer/endpoint/customer.endpoint";
-import {BaseApiAdapter} from "@utility/adapter/base.api.adapter";
-import {TypeGuard} from "@p4ck493/ts-type-guard";
-import {is} from "@utility/checker";
-import {HttpContext} from '@angular/common/http';
-import {TokensHttpContext} from '@src/tokens.http-context';
-
-type ResponseType = {
-  deletedCount: number
-};
-
-@Injectable({
-  providedIn: 'root'
-})
-export class RemoveCustomerApiAdapter extends BaseApiAdapter<ResponseType, [string]> {
-
-  /**
-   * DELETE ITEM BY ID
-   * @param id
-   */
-  @TypeGuard([is.string])
-  public override execute$(id: string) {
-    return this.httpClient.delete<ResponseType>(customerEndpointEnum.delete, {
-			context: new HttpContext().set(TokensHttpContext.REPLACE, {
-				id
-			}),
-    });
-  }
-
-}
