+++ src/module/customer/domain/entity/index.ts
@@ -1,0 +1,1 @@
+export * from './e.customer';
