+++ src/module/customer/domain/enum/index.ts
@@ -1,0 +1,1 @@
+export * from './customer-type.enum';;
