+++ src/module/customer/domain/index.ts
@@ -1,1 +1,3 @@
 export * from './interface';
+export * from './enum';
+export * from './entity';
