+++ src/module/customer/index.ts
@@ -1,0 +1,4 @@
+export * from './application';
+export * from './domain';
+export * from './infrastructure';
+export * from './presentation';
