+++ src/module/customer/infrastructure/api/index.ts
@@ -1,0 +1,7 @@
+export * from './archive.customer.api';
+export * from './create.customer.api';
+export * from './item.customer.api';
+export * from './list.customer.api';
+export * from './remove.customer.api';
+export * from './unarchive.customer.api';
+export * from './archive.customer.api';
