+++ src/module/customer/infrastructure/api/update.customer.api.ts
@@ -1,0 +1,25 @@
+import {Injectable} from '@angular/core';
+import {BaseApiAdapter} from "@utility/adapter/base.api.adapter";
+import {TypeGuard} from "@p4ck493/ts-type-guard";
+import {is} from "@utility/checker";
+import {HttpContext} from "@angular/common/http";
+import {TokensHttpContext} from "@src/tokens.http-context";
+import {customerEndpointEnum, ICustomer} from "@module/customer";
+
+@Injectable()
+export class UpdateCustomerApi extends BaseApiAdapter<ICustomer.DTO, [ICustomer.DTO]> {
+
+  /**
+   * SAVE NEW ITEM OR UPDATE ITEM BY ID
+   * @param value
+   */
+  @TypeGuard([is.object_not_empty])
+  public override execute$(value: ICustomer.DTO) {
+    return this.httpClient.put<ICustomer.DTO>(customerEndpointEnum.update, value, {
+			context: new HttpContext().set(TokensHttpContext.REPLACE, {
+				id: value._id,
+			}),
+    });
+  }
+
+}
