+++ src/module/customer/infrastructure/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * from './customer.endpoint';
