+++ src/module/customer/infrastructure/index.ts
@@ -1,0 +1,4 @@
+export * from './api';
+export * from './endpoint';
+export * from './repository';
+export * from './store';
