+++ src/module/customer/infrastructure/store/[case]/index.ts
@@ -1,0 +1,1 @@
+export * from './customer.main-page.store';
