+++ src/module/customer/infrastructure/store/index.ts
@@ -1,0 +1,4 @@
+export * from './customer.store';
+export * from './customer.presentation.store';
+export * from './[case]';
+
