--- src/module/customer/presentation/resolver/customer.cache.resolver.ts
@@ -1,16 +1,0 @@
-// export const customerCacheResolver: ResolveFn<boolean> = () => {
-//
-//   const store = inject(Store); // NGXS
-//
-//   const {app}: { app: IAppState } = store.snapshot();
-//
-//   if (app?.pageLoading) {
-//     return EMPTY;
-//   }
-//
-//   return store.dispatch(new CustomerActions.InitDefaultsFromCache())
-//     .pipe(
-//       map(() => true),
-//       catchError(() => EMPTY)
-//     );
-// };
