--- src/module/customer/state/customer/customer.actions.ts
@@ -1,4 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-namespace
-export namespace CustomerActions {
-
-}
